Moved prototypes to the header
